Add vendored dependency to the testify testing library via gvt
The vendor directory takes priority when golang looks for an import, therefore if the library was installed via `go get`, the one commited in the vendor directory will take priority.

This removes the need for manually getting the dependency during project setup.

This was installed via `gvt fetch` (https://github.com/FiloSottile/gvt), however it is not necessary to have gvt installed to build the project.